Begin a complete rewrite/rethink for v3.
This is an attempt to simplify the protocol and generally make
things easier to work with, as well as simplify the implementation
by utilizing engine.io for the transports.

This implementation tries to reduce the amount of state that
must be managed both by client and server. Most of the usage I
was seeing with Nydus would be clients CALLing methods on the
server, then immediately also SUBSCRIBEing to a relevant path.
Since the server would then have to do a permissions check on
both requests, and it probably already could have known the
client wanted to subscribe, that makes things unnecessarily
complex. As such, I've removed the SUBSCRIBE method altogether:
clients will now simply make CALLs to the server. If the server
wants to subscribe a client to a particular channel as the result
of a CALL (now called INVOKE to avoid Function.prototype.call
collision), it can do so. It can also unsubscribe the client at
any time without needing to notify them (but could via a publish
or a server->client INVOKE if it wanted to).

The implementation is not fully here as of this commit, still
very much a work in progress, but it actually has tests this
time! I'm committing all the pieces (protocol, server, client)
here for now, and will split them to their respective repos once
things become more stable.